@@ -2,62 +2,66 @@
 
 namespace InCube.Core.Diagnostic
 {
-    public interface IStopwatch
-    {
-        void Start();
-
-        void Stop();
-        
-        void Reset();
-        
-        void Restart();
-        
-        bool IsRunning { get; }
-        
-        TimeSpan Elapsed { get; }
-        
-        long ElapsedMilliseconds { get; }
-        
-        long ElapsedTicks { get; }
-    }
-
+    /// <summary>
+    /// Stopwatch that does nothing.
+    /// </summary>
     public sealed class DummyStopwatch : IStopwatch
     {
-        public static readonly DummyStopwatch Instance = new DummyStopwatch();
+        /// <summary>
+        /// The singleton <see cref="DummyStopwatch"/>.
+        /// </summary>
+        public static readonly DummyStopwatch Instance = new();
 
         private const int ZeroTicks = 0;
-        private static readonly TimeSpan EmptyTimeSpan = new TimeSpan(ZeroTicks);
+
+        private static readonly TimeSpan EmptyTimeSpan = new(ZeroTicks);
 
         private DummyStopwatch()
         {
         }
 
+        /// <inheritdoc/>
+        public bool IsRunning => false;
+
+        /// <inheritdoc/>
+        public TimeSpan Elapsed => EmptyTimeSpan;
+
+        /// <inheritdoc/>
+        public long ElapsedMilliseconds => ZeroTicks;
+
+        /// <inheritdoc/>
+        public long ElapsedTicks => ZeroTicks;
+
+        /// <summary>
+        /// Gets a reference to the single instance of the <see cref="DummyStopwatch"/>.
+        /// </summary>
+        /// <returns>The <see cref="DummyStopwatch"/>.</returns>
+        public static DummyStopwatch StartNew() => Instance;
+
+        /// <summary>
+        /// Gets zero.
+        /// </summary>
+        /// <returns>Zero.</returns>
+        public static long GetTimestamp() => ZeroTicks;
+
+        /// <inheritdoc/>
         public void Start()
         {
         }
 
-        public static DummyStopwatch StartNew() => Instance;
-
+        /// <inheritdoc/>
         public void Stop()
         {
         }
 
+        /// <inheritdoc/>
         public void Reset()
         {
         }
 
+        /// <inheritdoc/>
         public void Restart()
         {
         }
-
-        public bool IsRunning => false;
-
-        public TimeSpan Elapsed => EmptyTimeSpan;
-
-        public long ElapsedMilliseconds => ZeroTicks;
-
-        public long ElapsedTicks => ZeroTicks;
-
-        public static long GetTimestamp() => ZeroTicks;
     }
 }
