@@ -1,52 +1,41 @@
 ﻿using System;
 
-#pragma warning disable SA1400 // Access modifier should be declared
-
 namespace InCube.Core.Demo.Functional
 {
-    static class AnyDemo
+    internal static class AnyDemo
     {
-        readonly struct Any<T>
+        public readonly struct Any<T>
         {
-            public Any(T value)
-            {
-                Value = value;
-            }
+            public Any(T value) => this.Value = value;
 
             public T Value { get; }
 
-            public static implicit operator Any<T>(T value) => new Any<T>(value);
+            public static implicit operator Any<T>(T value) => new (value);
 
             public static implicit operator T(Any<T> any) => any.Value;
         }
 
-        readonly struct Option<T>
+        public readonly struct Option<T>
         {
+            public static readonly Option<T> None = default;
+
             private readonly Any<T>? value;
 
-            private Option(T value)
-            {
-                this.value = value;
-            }
+            private Option(T value) => this.value = value;
 
-            public static readonly Option<T> None = default(Option<T>);
+            public bool HasValue => this.value.HasValue;
 
-            public static implicit operator Option<T>(T value) =>
-                typeof(T).IsValueType || !ReferenceEquals(value, null) ? new Option<T>(value) : None;
+            public T Value => this.value.Value;
+
+            public static implicit operator Option<T>(T value) => typeof(T).IsValueType || !ReferenceEquals(value, null) ? new Option<T>(value) : None;
 
             public static explicit operator T(Option<T> option) => option.Value;
 
-            public bool HasValue => value.HasValue;
+            public T GetValueOrDefault() => this.value.GetValueOrDefault();
 
-            public T Value => value.Value;
+            public T GetValueOrDefault(T @default) => this.value.GetValueOrDefault(@default);
 
-            public T GetValueOrDefault() => value.GetValueOrDefault();
-
-            public T GetValueOrDefault(T @default) => value.GetValueOrDefault(@default);
-
-            public T GetValueOr(Func<T> @default) => value ?? @default.Invoke();
-
-            // ...
+            public T GetValueOr(Func<T> @default) => this.value ?? @default.Invoke();
         }
     }
 }
