@@ -1,7 +1,7 @@
 ﻿namespace InCube.Core.Functional
 {
     /// <summary>
-    /// Represents the Bottom type, i.e., the inner type of <see cref="Option.None"/> and <see cref="Maybe.None"/>.
+    /// Represents the Bottom type, i.e., the inner type of <see cref="Option.None"/> and <see cref="Maybes.None"/>.
     /// </summary>
     // ReSharper disable once ConvertToStaticClass
     // ReSharper disable once ClassNeverInstantiated.Global
