@@ -1,38 +1,28 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
-using System.Linq;
-using InCube.Core.Collections;
-using InCube.Core.Functional;
-using static InCube.Core.Preconditions;
+using JetBrains.Annotations;
 
 namespace InCube.Core.Numerics
 {
-
     /// <summary>
     /// Represents a histogram type.
     /// </summary>
-    /// <typeparam name="T"></typeparam>
-    [SuppressMessage("StyleCop.CSharp.MaintainabilityRules",
-        "SA1402:FileMayOnlyContainASingleType",
-        Justification = "Companion object.")]
+    /// <typeparam name="T">The type of objects in the histogram.</typeparam>
+    [SuppressMessage("StyleCop.CSharp.MaintainabilityRules", "SA1402:FileMayOnlyContainASingleType", Justification = "Companion object.")]
+    [PublicAPI]
     public class Histogram<T>
     {
-        internal Histogram(IReadOnlyList<T> binLowerEdges,
+        internal Histogram(
+            IReadOnlyList<T> binLowerEdges,
             IReadOnlyList<int> binCounts,
             IReadOnlyList<int> binIndices,
             int discardedCount,
             bool lowerEdgeEquals,
             Func<T, int> binSelector)
         {
-            CheckArgument(binLowerEdges.Count == binCounts.Count,
-                "number of bin lower edges {0} must match the number of bin count {1}",
-                binLowerEdges.Count,
-                binCounts.Count);
-            CheckArgument(discardedCount <= binIndices.Count,
-                "number of discarded elements {0} must be at most the total number of elements {1}",
-                discardedCount,
-                binIndices.Count);
+            Preconditions.CheckArgument(binLowerEdges.Count == binCounts.Count, "number of bin lower edges {0} must match the number of bin count {1}", binLowerEdges.Count, binCounts.Count);
+            Preconditions.CheckArgument(discardedCount <= binIndices.Count, "number of discarded elements {0} must be at most the total number of elements {1}", discardedCount, binIndices.Count);
             this.BinLowerEdges = binLowerEdges;
             this.BinCounts = binCounts;
             this.BinIndices = binIndices;
@@ -41,74 +31,43 @@ namespace InCube.Core.Numerics
             this.BinSelector = binSelector;
         }
 
+        /// <summary>
+        /// Gets the lower edges of the bins.
+        /// </summary>
         public IReadOnlyList<T> BinLowerEdges { get; }
-        
+
+        /// <summary>
+        /// Gets the count of elements per bin.
+        /// </summary>
         public IReadOnlyList<int> BinCounts { get; }
-        
+
+        /// <summary>
+        /// Gets the indices of the bins.
+        /// </summary>
         public IReadOnlyList<int> BinIndices { get; }
-        
+
+        /// <summary>
+        /// Gets the count of discarded elements.
+        /// </summary>
         public int DiscardedCount { get; }
-        
+
+        /// <summary>
+        /// Gets the total count.
+        /// </summary>
         public int TotalCount => this.BinIndices.Count;
-        
+
+        /// <summary>
+        /// Gets a value indicating whether the lower edge are equals.
+        /// </summary>
         public bool LowerEdgeEquals { get; }
 
         private Func<T, int> BinSelector { get; }
 
+        /// <summary>
+        /// Gets the bin in which the <paramref name="value"/> is located.
+        /// </summary>
+        /// <param name="value">The value to look for.</param>
+        /// <returns>The index of the bin in which <paramref name="value"/> is located.</returns>
         public int FindBin(T value) => this.BinSelector(value);
-    }
-
-    public static class Histogram
-    {
-        public const bool DefaultLowerEdgeEquals = true;
-
-        public static Histogram<T> Create<T>(IEnumerable<T> values,
-            IReadOnlyList<T> edges,
-            IComparer<T> comparer,
-            bool lowerEdgeEquals = DefaultLowerEdgeEquals)
-        {
-            CheckArgument(edges.IsSorted(comparer: comparer), "edges must be sorted");
-
-            var binCount = edges.Count;
-            var binSelector = Functions.Invoke(() =>
-            {
-                if (edges is T[] edgeArray)
-                {
-                    return lowerEdgeEquals
-                        ? (Func<T, int>)(v => edgeArray.UpperBound(v, comparer) - 1)
-                        : (Func<T, int>)(v => edgeArray.LowerBound(v, comparer) - 1);
-                }
-
-                var listEdges = edges as List<T> ?? edges.ToList();
-                return lowerEdgeEquals
-                    ? (Func<T, int>)(v => listEdges.UpperBound(v, comparer) - 1)
-                    : (Func<T, int>)(v => listEdges.LowerBound(v, comparer) - 1);
-            });
-
-            var (counts, indices, discarded) = GenericHistogram(values, binCount, binSelector);
-            return new Histogram<T>(edges, counts, indices, discarded, lowerEdgeEquals, binSelector);
-        }
-
-        private static (int[] Counts, List<int> Indices, int Discarded) GenericHistogram<T>(
-            IEnumerable<T> values,
-            int binCount,
-            Func<T, int> binSelector)
-        {
-            var indices = values.Select(binSelector).ToList();
-            var counts = new int[binCount];
-            var discarded = 0;
-            foreach (var index in indices)
-            {
-                if (index >= 0)
-                {
-                    ++counts[index];
-                }
-                else
-                {
-                    ++discarded;
-                }
-            }
-            return (counts, indices, discarded);
-        }
     }
 }
