@@ -8,9 +8,15 @@ namespace InCube.Core.Numerics
     /// <summary>
     /// A range type representing a closed interval.
     /// </summary>
-    /// <typeparam name="T"></typeparam>
-    public readonly struct Range<T> : IInvariantRange<T, Range<T>> where T : IComparable<T>
+    /// <typeparam name="T">The type of the lower and upper bound.</typeparam>
+    public readonly struct Range<T> : IInvariantRange<T, Range<T>>
+        where T : IComparable<T>
     {
+        /// <summary>
+        /// Initializes a new instance of the <see cref="Range{T}"/> struct.
+        /// </summary>
+        /// <param name="min">The lower bound of the range.</param>
+        /// <param name="max">The upper bound of the range.</param>
         [JsonConstructor]
         public Range(T min, T max)
         {
@@ -19,27 +25,33 @@ namespace InCube.Core.Numerics
             this.Max = max;
         }
 
-        public Range<T> With(Option<T> min = default, Option<T> max = default) =>
-            new Range<T>(min.GetValueOrDefault(this.Min), max.GetValueOrDefault(this.Max));
+        public Range<T> With(Option<T> min = default, Option<T> max = default) => new(min.GetValueOrDefault(this.Min), max.GetValueOrDefault(this.Max));
 
+        /// <inheritdoc/>
         public T Min { get; }
 
+        /// <inheritdoc/>
         public T Max { get; }
 
+        public static bool operator ==(Range<T> left, Range<T> right) => Equals(left, right);
+
+        public static bool operator !=(Range<T> left, Range<T> right) => !(left == right);
+
+        /// <inheritdoc/>
         public bool Contains(T x) => this.Min.CompareTo(x) <= 0 && x.CompareTo(this.Max) <= 0;
 
-        public bool Contains(Range<T> that) => 
-            this.Min.CompareTo(that.Min) <= 0 && that.Max.CompareTo(this.Max) <= 0;
+        /// <inheritdoc/>
+        public bool Contains(Range<T> range) => this.Min.CompareTo(range.Min) <= 0 && range.Max.CompareTo(this.Max) <= 0;
 
-        public bool OverlapsWith(Range<T> that) =>
-            this.Min.CompareTo(that.Max) <= 0 && that.Min.CompareTo(this.Max) <= 0;
+        /// <inheritdoc/>
+        public bool OverlapsWith(Range<T> range) => this.Min.CompareTo(range.Max) <= 0 && range.Min.CompareTo(this.Max) <= 0;
 
-        public Range<T> IntersectWith(Range<T> that) => 
-            Comparables.Max(this.Min, that.Min).ToRange(Comparables.Min(this.Max, that.Max));
+        /// <inheritdoc/>
+        public Range<T> IntersectWith(Range<T> range) => Comparables.Max(this.Min, range.Min).ToRange(Comparables.Min(this.Max, range.Max));
 
-        #pragma warning disable CA2225 // Operator overloads have named alternates
+#pragma warning disable CA2225 // Operator overloads have named alternates
         public static implicit operator Range<T>((T Min, T Max) that) =>
-            #pragma warning restore CA2225 // Operator overloads have named alternates
+#pragma warning restore CA2225 // Operator overloads have named alternates
             that.Min.ToRange(that.Max);
 
         public void Deconstruct(out T min, out T max)
@@ -48,19 +60,19 @@ namespace InCube.Core.Numerics
             max = this.Max;
         }
 
-        public bool Equals(Range<T> that) => 
-            this.Min.CompareTo(that.Min) == 0 && this.Max.CompareTo(that.Max) == 0;
+        /// <inheritdoc/>
+        public bool Equals(Range<T> that) => this.Min.CompareTo(that.Min) == 0 && this.Max.CompareTo(that.Max) == 0;
 
-        public override bool Equals(object obj)
+        /// <inheritdoc/>
+        public override bool Equals(object? obj)
         {
             if (ReferenceEquals(null, obj))
-            {
                 return false;
-            }
 
             return obj is Range<T> range && this.Equals(range);
         }
 
+        /// <inheritdoc/>
         public override int GetHashCode()
         {
             unchecked
@@ -69,22 +81,6 @@ namespace InCube.Core.Numerics
             }
         }
 
-        public static bool operator ==(Range<T> left, Range<T> right) => Equals(left, right);
-
-        public static bool operator !=(Range<T> left, Range<T> right) => !(left == right);
-
         public override string ToString() => $"[{this.Min}, {this.Max}]";
-    }
-
-    public static class Range
-    {
-        public static Range<T> ToRange<T>(this T min, T max) where T : IComparable<T> =>
-            new Range<T>(min, max);
-
-        public static Range<T> ToRange<T>(this IRange<T> range) where T : IComparable<T> =>
-            range.Min.ToRange(range.Max);
-
-        [Obsolete("unnecessary call")]
-        public static Range<T> ToRange<T>(this Range<T> range) where T : IComparable<T> => range;
     }
 }
