@@ -44,8 +44,8 @@ namespace InCube.Core.Test.Format
         public void TestGenericNullRefConverter()
         {
             var converter = new GenericOptionJsonConverter(typeof(Maybe<>));
-            Assert.True(converter.CanConvert(Maybe.None.GetType()));
-            TestOptionsConverter(converter, Maybe.Some(Boxed.Of(42)));
+            Assert.True(converter.CanConvert(Maybes.None.GetType()));
+            TestOptionsConverter(converter, Maybes.Some(Boxed.Of(42)));
         }
 
         [Test]
